Update the grammar to accept non-hex references (#49)
This in turn, caused a different issue and that is that ANTLR was
considering CollectionViewController a REFERENCE (because it's 24
characters long and the lexer rule appears before NON_QUOTED_STRING).

I fixed it by allowing only upper case letters in references but this is
more a patch than a fix. Now that Carthage is breaking the only-hex
characters for references, we have it way harder to differentiate those
from just strings in the lexer phase, which is a pain. I hope we can
live with that for some more time, though.

I made the grammar more restrictive regarding numeric values as well.