Reorder some rules
The order of rules in ANTLR is quite important, as they take precedence
based on when they are defined. Here, I'm putting the ALPHA_NUMERIC and
ALPHA_NUMERIC_CAP rules below the QUOTED_STRING and NON_QUOTED_STRING
rules so that when we find a single-letter identifier (i.e. a file name)
ANTLR detects the *_STRING rules before the ALPHA_NUMERIC_* ones.

This is most likely caused by a badly defined grammar (and probably the
lack of proper fragments). However, this is, I think, a good compromise
if we don't want to tackle a full rewrite of the grammar.